actions: Call send_initial_pms from process_new_human_user.
